SVN-Revision: 5219

+++ src/squirrel/sqstdlib/sqstdstream.cpp
@@ -163,7 +163,7 @@ SQInteger _stream_writen(HSQUIRRELVM v)
 	case 'f': {
 		float f;
 		sq_getfloat(v, 2, &tf);
-		f = tf;
+		f = (float)tf;
 		self->Write(&f, sizeof(float));
 			  }
 		break;
@@ -302,7 +302,7 @@ SQRESULT declare_stream(HSQUIRRELVM v,SQChar* name,SQUserPointer typetag,SQChar*
 		}
 		sq_createslot(v,-3);
 		sq_pop(v,1);
-
+		
 		i = 0;
 		while(globals[i].name!=0)
 		{
